Add some PY2/PY3 compatibility functions
- pyformex.round(number,ndigits): performs Python3 rounding behavior
Values ending on 5 are rounded to the nearest even number.
It is recommended to add 'from pyformex import round' in all code
where you use the builtin round function. This will result in the
same behavior under py2 and py3. Note that
   round(number)  returns an int
   round(number, 0)  returns a float with 0 fraction
   round(number, ndigits) returns a float rounded to ndigits
Note that numpy round/around has the same behavior, except for
the int return when ndigits is omitted.

- pyformex.isFile and pyformex.isString can be used to test whether
an object is a file, or a string (ascii/bytes or unicode) type.

+++ pyformex/gui/viewport.py
@@ -31,7 +31,7 @@ implements a dynamic array of multiple canvases.
 from __future__ import absolute_import, division, print_function
 
 import pyformex as pf
-from pyformex import zip, utils
+from pyformex import zip, round, utils
 
 from pyformex.opengl import canvas
 from pyformex.gui import (
@@ -434,9 +434,9 @@ class QtCanvas(QtOpenGL.QGLWidget, canvas.Canvas):
         if width<=0 or height <= 0:
             wc,hc = self.getSize()
             if height > 0:
-                width = int(round(float(height)/hc*wc))
+                width = round(float(height)/hc*wc)
             elif width > 0:
-                height = int(round(float(width)/wc*hc))
+                height = round(float(width)/wc*hc)
             else:
                 width,height = wc,hc
         return width, height
